[WIP] Started working on fixing compensations with new inputs-based method. + Removed some obsolete part of the code.
+ Added a "limit" for the update. The loop will run every 10ms to avoid cluster and it grabbing all resources.

+ Optimized some of the code for faster execution and smaller packets.
> Position now is a short between -100 and 100 instead of a float. This makes X and Y pos. 2x smaller than normal.
> Inputs aren't added on server if they are duplicate from last one.

+++ communication/Input.hpp
@@ -10,8 +10,8 @@ class Input {
 private:
     // TODO: unsigned int limit is at 4,294,967,295. This means it should be reset to 0 on round start!
     unsigned int id;
-    float movementX;
-    float movementY;
+    short movementX;
+    short movementY;
     float rotate;
     bool onController;
     bool changeWpn;             // Signal to change weapon
